[headless] Initial attempt at a headless, cmdline mode for MakeHuman
MakeHuman is a powerful tool to procedurally generate realistic human models.
To be useful as input source for other tools (like crowd simulators, human-robot
interaction simulators, etc), a command-line interface is desirable.

This commit is *a first step* towards this goal, with currently very limited
features. If it get positive feedback, we can easily extend it to cover more of
MakeHuman features.

To use it, simply pass the option '--output=myexportedmodel.mhx'.

Current features:

- console mode: do no require OpenGL/X connection
- allow to parameter a few options like the gender (see makehuman --help)
- support addition of rigs, hairs, body textures
- export the model to MHX